Added sample endpoint that allows to retrive latest measure.
In order to get deeper knowledge about the persistenEntityRegistry,
it was tested how to recover some entity given some id.

In order to recover all posible events, envetStream can be read, but,
it's not suitable for temporal series.

OTOH, using more complex entityId that includes information for both
device Id and temporal information, neither solve the problem of
requesting some temporal serie.

So that, in order to have some example of how to use the
persistenEntityRegistry for recover entity values, it was included some
endpoint that allows to read latest added measure.