Writing flatten version of measurement on Cassandra.
In order to deal with tempora series, it's better using some flatten
version of measurement, thus, they must be stored into some cassandra
table, in order to be later consulted.

First version of read processor was included, but test where not
included yet, due to it's waiting to include read end-points.